# sdk/ts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vmond",
    "description": "Vibemon server core: engine, image pipeline, v1 HTTP API, mesh",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/sandboxes/{id}/files/stat": {
      "get": {
        "tags": [],
        "operationId": "file_stat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/v1/sandboxes/{id}/migrate": {
      "post": {
        "tags": [],
        "operationId": "migrate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MigrateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MigrateBody": {
        "type": "object",
        "description": "`POST /{id}/migrate` body (admin; mesh).",
        "required": [
          "target"
        ],
        "properties": {
          "target": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}